requirements.txt
transformers>=4.44.0
torch>=2.0.0
pillow>=10.0.0
numpy>=1.24.0
accelerate>=0.25.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface_hub>=0.20.0
hf_xet>=0.3.0
requests>=2.28.0
flask>=2.3.0

